Move cache key functions to a separate module
